fix: correct failover semantics, protocol handling, and parsing; extend tests
Review findings fixed:
- Application-level "failed" answers no longer count as transport errors:
  new ResponseError type, no wrong HA failover, and RegisterHost's
  documented failed-then-confirm autoregistration flow now actually works
  (was dead code).
- UpdateHost is now implemented: caches the final redirect target as
  PrimaryHost (was declared and documented but never read).
- GetInfo returns parse errors for processed/failed/total instead of
  silently reporting zeros.
- normalizeHost handles IPv6 literals ("::1" -> "[::1]:10051").
- Data length written as uint64 (was PutUint32 into the 8-byte field).
- Response datalen header is validated instead of ignored.
- Send wraps the last underlying error instead of discarding it, and an
  empty host list reports "no hosts configured".
- Packet is marshaled once per send; marshal errors are checked.
- Constructors deduplicated; RegisterHost simplified; README typos,
  missing time import, and Go version (1.20+) corrected.

Tests: new suites for redirects, UpdateHost, failover, cache
invalidation, terminal failed answers, timeouts, truncated/oversized
responses, GetInfo/normalizeHost/parseHostPort tables, wire format, and
the real autoregistration flow. Real-server integration tests moved
behind -tags integration, configured via ZABBIX_ADDR instead of
hardcoded localhost. Coverage ~51% -> 95.7%, all green under -race.

+++ packet.go
@@ -37,6 +37,6 @@ func NewPacket(data []*Metric, agentActive bool, t ...time.Time) *Packet {
 func (p *Packet) DataLen() []byte {
 	dataLen := make([]byte, 8)
 	JSONData, _ := json.Marshal(p)
-	binary.LittleEndian.PutUint32(dataLen, uint32(len(JSONData)))
+	binary.LittleEndian.PutUint64(dataLen, uint64(len(JSONData)))
 	return dataLen
 }
